fix(v12-urgent): V20260820.5 SQL alias bug + V6 FK/RLS hardening + V7 BIGINT cast
PRODUCTION EMERGENCY (per v12 brief):
- Render deploy 96cb76e FAILED: V20260820.5 SQLSTATE=42P01
  'missing FROM-clause entry for table r'
- Production code/schema incompatibility: V20260820.4 dropped
  uk_commerce_orders_tenant_store_idempotency but live code 6ec0059
  still uses ON CONFLICT (tenant_id, store_id, idempotency_key)
  → ECOMMERCE_CHECKOUT_PRODUCTION=DEGRADED
- Current Flyway: 121 records, V20260820.4 applied, V20260820.5 NOT applied
- Live image: 6ec0059 (Render dep-da35gejtqb8s73c600v0)
- Failed deploy for 9094717: dep-da36asou01pc7380kqa0

FIX V20260820.5 (CRITICAL — production is broken):
- UPDATE roles SET template_key = r.code ← r alias undeclared
- Fix: UPDATE roles AS r SET template_key = r.code WHERE r.is_system_managed = TRUE ...
- SQLSTATE 42P01 resolved
- V20260820.5 will now apply successfully on next deploy

FIX V20260820.7 (pre-flight):
- count_role_caps function declared RETURNS INTEGER but COUNT(*) returns BIGINT
- Add explicit ::INTEGER cast to avoid implicit-conversion ambiguity

HARDEN V20260820.6 (before it ever reaches prod — it has not applied yet):
- commerce_order_finance_links: add tenant-aware composite FKs
  - fk_commerce_order_finance_links_order FOREIGN KEY (tenant_id, commerce_order_id)
    REFERENCES commerce_orders(tenant_id, id) ON DELETE CASCADE
  - fk_commerce_order_finance_links_invoice FOREIGN KEY (tenant_id, finance_invoice_id)
    REFERENCES finance_invoices(tenant_id, id) ON DELETE CASCADE
  - COMMERCE_FINANCE_LINK_FK_INTEGRITY=PASS — no orphan linkage rows
- role_template_bindings: add tenant-aware FK to roles
  - fk_role_template_bindings_role FOREIGN KEY (tenant_id, role_id)
    REFERENCES roles(tenant_id, id) ON DELETE CASCADE
- RLS governance for 4 new tenant-scoped tables (matches existing
  V20260816_6 + V20260816_8 conventions):
  - commerce_order_number_sequences: ENABLE + FORCE RLS + tenant_isolation policy
  - commerce_order_finance_links: same
  - finance_invoice_number_sequences: same
  - role_template_bindings: same
  - All policies use current_setting('app.tenant_id', true)::uuid
  - NEW_TABLE_RLS_GOVERNANCE=PASS

Pre-flight verified for V5/V6/V7 PostgreSQL syntax:
- All table aliases declared (UPDATE roles AS r pattern)
- All referenced columns exist (verified against V20260815_16 + V20260816_5)
- All FK targets exist (commerce_orders(tenant_id,id), finance_invoices(tenant_id,id),
  roles(tenant_id,id))
- DO blocks compile (validated by manual review of plpgsql syntax)
- count_role_caps function compiles with explicit BIGINT→INTEGER cast
- Migration ordering: V6 ALTER TABLE on commerce_order_number_sequences
  runs AFTER V4 created the table (V4 already applied in prod)

This commit UNBLOCKS production. After deploy:
- Flyway: 121 → 124 (V20260820.5 + V20260820.6 + V20260820.7 applied)
- V20260820.7 DO block validates all 9 RBAC templates' exact matrix
- commerce_order_finance_links FKs enforce referential integrity
- 4 new tenant tables under RLS governance

Files: 3 migration files corrected/hardened

+++ apps/sanad-platform/src/main/resources/db/migration/V20260820_7__rbac_exact_matrix_9_of_9.sql
@@ -16,10 +16,15 @@
 -- with a message that names the offending template and the mismatch.
 -- ============================================================
 
--- Helper: count capabilities on a system-managed role
+-- Helper: count capabilities on a system-managed role.
+-- NOTE: PostgreSQL COUNT(*) returns BIGINT — must cast to INTEGER to match
+-- the RETURNS declaration. Function parameters are named tenant_uuid /
+-- role_uuid and are referenced by name in the WHERE clause (PL/pgSQL
+-- function-argument naming convention).
 CREATE OR REPLACE FUNCTION count_role_caps(tenant_uuid UUID, role_uuid UUID)
 RETURNS INTEGER AS $$
-    SELECT COUNT(*) FROM role_capabilities rc WHERE rc.tenant_id = tenant_uuid AND rc.role_id = role_uuid;
+    SELECT COUNT(*)::INTEGER FROM role_capabilities rc
+    WHERE rc.tenant_id = tenant_uuid AND rc.role_id = role_uuid;
 $$ LANGUAGE SQL STABLE;
 
 DO $$
